backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv
pydantic
pydantic-settings

sqlalchemy
aiosqlite

openai
tiktoken

sentence-transformers
faiss-cpu
numpy

pymupdf
python-docx

streamlit
requests